renamed package from mcr to icr, fixed integer bug in covid example scm
