+++ .gitignore
@@ -1,5 +1,5 @@
 # my choices
-jobs/
+icr-jobs/
 launch.json
 results.zip
 code-download.zip
