+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
-name = mcr-experiments
-version = 0.0.1
+name = icr-experiments
+version = 0.0.2
 author = XXXX
 author_email = XXXX
 description = Python package for experiments for the Meaningful Algorithmic Recourse paper.
