+++ src/icr/causality/dags.py
@@ -3,7 +3,7 @@ from typing import List
 import matplotlib.pyplot as plt
 import networkx as nx
 
-from mcr.utils import search_nonsorted
+from icr.utils import search_nonsorted
 
 
 class DirectedAcyclicGraph:
