+++ src/icr/experiment/__init__.py
@@ -1,0 +1,2 @@
+from icr.experiment.compile import compile_experiments
+from icr.experiment.run import run_experiment
