--- src/mcr/causality/__init__.py
@@ -1,2 +1,0 @@
-from mcr.causality.dags import DirectedAcyclicGraph
-from mcr.causality.scms.scm import *
