--- src/mcr/experiment/__init__.py
@@ -1,2 +1,0 @@
-from mcr.experiment.compile import compile_experiments
-from mcr.experiment.run import run_experiment
