--- src/mcr/recourse/__init__.py
@@ -1,3 +1,0 @@
-from mcr.recourse.recourse import recourse
-from mcr.recourse.population import recourse_population
-from mcr.recourse.utils import save_recourse_result, compute_h_post_individualized
